requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Machine Learning
torch==2.1.0
torchaudio==2.1.0
transformers==4.36.0
openai-whisper==20231117
scikit-learn==1.3.2
numpy==1.24.3
nltk==3.8.1
fasttext-wheel==0.9.2

# Audio Processing
librosa==0.10.1
pydub==0.25.1
soundfile==0.12.1
python-speech-features==0.6

# System Monitoring & Performance
psutil==5.9.6

# Async & Utilities
aiofiles==23.2.1
redis==5.0.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1

# Environment
python-dotenv==1.0.0